requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pandas
openpyxl
pymupdf
gspread
google-auth
cloudinary
